refactor(cli): extract gate policy loading
