+++ docs/architecture-consolidation-plan.md
@@ -68,7 +68,7 @@ fixtures:
 | Analysis complexity | `crates/tokmd-analysis/src/complexity/mod.rs` + `complexity/functions.rs` + `complexity/functions/` + `complexity/details.rs` + `complexity/details/` + `complexity/summary.rs` + `complexity/risk.rs` + `complexity/debt.rs` + `complexity/histogram.rs` + `complexity/language.rs` + `complexity/math.rs` + `complexity/tests/unit.rs` | 156 + 223 + Rust function owner 126 + 122 + Rust span owner 37 + Python span owner 79 + JS/TS span owner 57 + Go span owner 47 + C-family span owner 52 + 138 + 78 + 69 + 33 + 35 + 5 + 346 | Keep shared complexity logic in `tokmd-analysis`, with Rust function counting plus Rust, Python, JavaScript/TypeScript, Go, and C-family function-span detection in owner modules; split remaining language/source/summary helpers and local unit tests only when fresh evidence shows a clear next seam |
 | CLI parser | `crates/tokmd/src/cli/parser.rs` and `crates/tokmd/src/cli/parser/` | `parser.rs` 209; command enum owner 103; global scan args owner 133; shared value-enum owner 235; analyze parser owner 178; context/handoff parser owner 166; cockpit/baseline parser owner 89; diff parser owner 67; gate parser owner 55; sensor parser owner 43; export parser owner 43; badge parser owner 45; module parser owner 36; lang parser owner 26; run parser owner 25; check-ignore parser owner 15; completions parser owner 21; init parser owner 36; tools parser owner 15 | Context, handoff, analyze, cockpit, baseline, diff, gate, sensor, export, badge, module, lang, run, check-ignore, completions, init, tools, command enum, global scan args, and shared value enums now live under parser owner modules; continue command-family splits only when clap snapshots prove behavior is unchanged |
 | CLI config resolution | `crates/tokmd/src/config.rs` and `crates/tokmd/src/config/resolve.rs` | `config.rs` 313; resolver owner 456 | Keep config discovery, profile selection, `ConfigContext`, and `ResolvedConfig` in the root config module while lang/module/export CLI-to-receipt argument resolution lives in the resolver owner module. Preserve root re-exports. |
-| Gate command | `crates/tokmd/src/commands/gate.rs` and `crates/tokmd/src/commands/gate/` | `gate.rs` 396; render owner 111 | Gate policy loading, ratchet loading, and receipt computation remain in the command coordinator while text/JSON result rendering lives in a narrow owner module. Keep `tokmd_gate` proof scoped to both command paths. |
+| Gate command | `crates/tokmd/src/commands/gate.rs` and `crates/tokmd/src/commands/gate/` | `gate.rs` 229; policy owner/tests 215; render owner 111 | Gate policy loading, ratchet loading, and config-rule conversion live in the policy owner; text/JSON result rendering lives in the render owner; receipt computation remains in the command coordinator. Keep `tokmd_gate` proof scoped to all gate command paths. |
 | Sensor command | `crates/tokmd/src/commands/sensor.rs` and `crates/tokmd/src/commands/sensor/` | `sensor.rs` 286; findings owner/tests 344; gate mapping owner/tests 189 | Sensor finding emission and cockpit-evidence-to-envelope gate mapping now live in owner modules while the command keeps cockpit execution and the 3-layer output topology; split rendering or output only when fresh evidence shows the command coordinator is still too broad |
 | Model aggregation | `crates/tokmd-model/src/lib.rs`, `crates/tokmd-model/src/aggregate.rs`, `crates/tokmd-model/src/rows.rs`, and `crates/tokmd-model/src/sorting.rs` | `lib.rs` 267; aggregation owner/tests 427; row collection owner/tests 411; sorting owner/tests 91 | Report builders, file-row collection, in-memory row detection, and row sorting now live in owner modules; remaining work is child-language behavior only if future evidence shows the seam is still too broad |
